handlers

+++ store/pg/migrations/20200629153446_init.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE clients (
 	id uuid NOT NULL,
 	bio text NOT NULL,
-	balance money NOT NULL,
+	balance float NOT NULL,
 	created_at timestamp default current_timestamp,
 	CONSTRAINT "pk_client_id" PRIMARY KEY (id)
 );
